requirements.txt
coverage==6.3.2
Django==4.0.3
django-extensions==3.1.5
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
psycopg2-binary==2.9.3
requests==2.27.1
